:art: Move helper functions into named modules
